+++ src/system/delay_actor.rs
@@ -1,5 +1,6 @@
 use std::thread::sleep;
 use std::time::Duration;
+use log::debug;
 use crate::message::delayed_message::DelayedMessage;
 use crate::prelude::{Actor, ActorContext, ActorFactory, ActorMessage, ActorResult, Handler};
 
@@ -31,11 +32,14 @@ impl<A, M> Handler<DelayedMessage<A, M>> for DelayActor
     fn handle(&mut self, msg: DelayedMessage<A, M>, context: &ActorContext<Self>) -> ActorResult {
         let duration = msg.started.elapsed();
         if duration >= msg.delay {
-            msg.destination.send(msg.msg);
+            let result = msg.destination.send(msg.msg);
+            if result.is_err() {
+                debug!("");
+            }
         }
         else {
             sleep(Duration::from_millis(100));
-            context.actor_ref.send(msg);
+            let _ = context.actor_ref.send(msg);
         }
 
         return ActorResult::Ok;