+++ CHANGELOG.md
@@ -3,6 +3,9 @@
  - actors can now delay message processing by going into a sleep state
    - can be achieved through `actor.sleep(duration)` on the `ActorWrapper` or by returning `ActorState::Sleep(duration)` from within the actor
  - fixed a bug where an actor with a limited mailbox can get stuck on shutdown
+ - added `send_timeout()`
+ - added return `Result<(), ActorSendError>` to message sending functions
+
 
 # 0.6.0